Initial commit of sensus
A tiddlyspace tool that uses css transforms and other frippery
to enable a view of recent happenings on tiddlyspace.

This is another tool to experiment with different css handling
of search results and websocket feeds, to eventually lead to some
useful library code.

Not done yet. Still need to present the tiddlers nicely. This code
makes no pretensions about being correct or pretty. There's lots
of dumb duplication, scoping ick, and other mess, but that will be
cleaned up eventually.

What it does so far:

* arrow keys move the center view left and right
* space bar goes to end of the list
* the future is to the right